Document: postSignUp 함수 Auth.ts로 이동
